+++ components/SSI/Stake/AddFunds/styles.module.scss
@@ -58,10 +58,7 @@
     color: #fff;
 }
 
-.zilpayWalletInfo {
-    display: flex;
-    cursor: pointer;
-    width: fit-content;
+.walletInfo {
     margin-bottom: 20px;
     margin-top: -10px;
 }
