+++ components/SSI/Stake/AddFunds/styleslight.module.scss
@@ -64,10 +64,7 @@
     color: #292828 !important;
 }
 
-.zilpayWalletInfo {
-    display: flex;
-    cursor: pointer;
-    width: fit-content;
+.walletInfo {
     margin-bottom: 20px;
     margin-top: -10px;
 }
