+++ components/SSI/Stake/Wallet/Dashboard/index.tsx
@@ -231,7 +231,7 @@ function DashboardStake({ balance }) {
                                 }
                             }
                         })
-                        .catch(() => { })
+                        .catch(() => {})
                 }
             }
         }
