+++ components/index.ts
@@ -69,3 +69,4 @@ export { default as WalletSelector } from './SSI/Stake/WalletSelector'
 export { default as SearchBarWallet } from './SSI/Stake/SearchBarWallet'
 export { default as Spinner } from './Spinner'
 export { default as ConnectButton } from './SSI/ConnectButton'
+export { default as WalletInfo } from './SSI/WalletInfo'
