+++ public/locales/es/common.json
@@ -337,6 +337,6 @@
     "UPDATE SOCIAL TREE": "ACTUALIZAR ÁRBOL SOCIAL",
     "Select source": "Seleccioná el origen de los fondos",
     "Type label": "Ingresá la etiqueta",
-    "Source of funds": "Source of funds",
-    "Select wallet": "Select wallet"
+    "Source of funds": "Origen de los fondos",
+    "Select wallet": "Seleccioná billetera"
 }
